requirements.txt
asgiref==3.2.3
backcall==0.1.0
beautifulsoup4==4.6.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cssselect2==0.2.2
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.8
django-braces==1.13.0
django-enumchoicefield==1.1.0
django-environ==0.4.5
django-hitcount==1.3.1
django-modelcluster==5.0
django-taggit==1.2.0
django-treebeard==4.3
django-unique-slugify==1.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
ffmpeg-python==0.2.0
future==0.18.2
html5lib==1.0.1
idna==2.8
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
l18n==2018.5
packaging==19.2
parso==0.5.2
pathlib==1.0.1
pdf2image==1.11.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
preview-generator==0.13
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
pycparser==2.19
pyexifinfo==0.4.0
Pygments==2.5.2
pyparsing==2.4.6
PyPDF2==1.26.0
python-magic==0.4.15
python-slugify==1.2.6
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tinycss2==1.0.2
traitlets==4.3.3
transifex-client==0.13.7
Unidecode==1.1.1
urllib3==1.25.7
vtk==8.1.2
wagtail==2.7
wagtailmedia==0.4.0
wagtailtrans==2.0.6
Wand==0.5.8
wcwidth==0.1.8
webencodings==0.5.1
Willow==1.3
xlrd==1.2.0
xvfbwrapper==0.2.9
